Fix magic method access of post
See https://github.com/mlaphp/mlaphp/issues/4

When accessing the post variable, I get a warning regarding accessing
the variable via magic method because the referenced field is not found
in the subject class. This is due to the variable being created in the
constructor and is not explicitly created like all the other
superglobals.

* Create public variable $post